perf: optimize editor rendering and shutdown
Port a set of low-risk editor optimizations:

- Grid: render all lines in a single glBegin/glEnd batch with one color
  set, and skip drawing the grid when zoomed out (zoom > 10) to cut overdraw.
- Close: after the save prompts, persist only the cheap state (hotkeys,
  perspective, recent files, client versions, settings), shut down
  process-COM services and exit immediately, avoiding the slow teardown
  that froze the app when closing large maps.
- Outfit: cast look components to uint32_t in getColorHash to avoid
  signed-shift undefined behavior.

- Build: remove dangling asio references left in CMake after the dependency was dropped.

+++ source/outfit.h
@@ -36,7 +36,7 @@ struct Outfit {
 	int lookMountFeet;
 
 	uint32_t getColorHash() const {
-		return lookHead << 24 | lookBody << 16 | lookLegs << 8 | lookFeet;
+		return static_cast<uint32_t>(lookHead) << 24 | static_cast<uint32_t>(lookBody) << 16 | static_cast<uint32_t>(lookLegs) << 8 | static_cast<uint32_t>(lookFeet);
 	}
 };
